requirements.txt
Django==3.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
mysql-client==0.0.1
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.17.0
requests==2.25.1
